+++ CHANGELOG.md
@@ -11,6 +11,7 @@ As of May 27, 2014
 * New
     - Application (replaces Project)
     - Acl (separated from Auth)
+    - Row Gateway and Table Gateway classes in the Db component
 * Revised
     - Auth
         + Separated the Acl component and moved to its own folder
@@ -19,7 +20,7 @@ As of May 27, 2014
     - Db
         + Removed the top-level Pop\Db\Db class in favor of direct access to the Db\Adapter classes
         + Removed Escaped adapter and support for non-prepared queries
-        + Revised teh join() method to be more clear in the Sql\Select class
+        + Revised the join() method to be more clear in the Sql\Select class
     - File
         + Completely stripped down to only the upload and checkDuplicate static methods
         + Trimmed Dir class, removed references to old File class